requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Image Processing
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Optional: For advanced analysis
scipy>=1.11.0
scikit-learn>=1.3.0
